+++ apps/web/src/app/billing/organizations/billing-sync-api-key.component.ts
@@ -7,8 +7,8 @@ import { PlatformUtilsService } from "@bitwarden/common/abstractions/platformUti
 import { UserVerificationService } from "@bitwarden/common/abstractions/userVerification/userVerification.service.abstraction";
 import { OrganizationApiServiceAbstraction } from "@bitwarden/common/admin-console/abstractions/organization/organization-api.service.abstraction";
 import { OrganizationApiKeyType } from "@bitwarden/common/admin-console/enums/organization-api-key-type";
+import { OrganizationApiKeyRequest } from "@bitwarden/common/admin-console/models/request/organization-api-key.request";
 import { ApiKeyResponse } from "@bitwarden/common/auth/models/response/api-key.response";
-import { OrganizationApiKeyRequest } from "@bitwarden/common/models/request/organization-api-key.request";
 import { Verification } from "@bitwarden/common/types/verification";
 
 export interface BillingSyncApiModalData {
