+++ libs/common/src/admin-console/models/request/organization-api-key.request.ts
@@ -1,5 +1,5 @@
-import { OrganizationApiKeyType } from "../../admin-console/enums/organization-api-key-type";
-import { SecretVerificationRequest } from "../../auth/models/request/secret-verification.request";
+import { SecretVerificationRequest } from "../../../auth/models/request/secret-verification.request";
+import { OrganizationApiKeyType } from "../../enums/organization-api-key-type";
 
 export class OrganizationApiKeyRequest extends SecretVerificationRequest {
   type: OrganizationApiKeyType = OrganizationApiKeyType.Default;
